add Android Kotlin: Forecast App 07 - Current Weather User Interface - MVVM Tutorial Course

+++ app/src/main/java/com/nakano/forecastmvvm/data/db/unitlocalized/MetricCurrentWeatherEntry.kt
@@ -4,19 +4,19 @@ import androidx.room.ColumnInfo
 
 data class MetricCurrentWeatherEntry(
     @ColumnInfo(name = "temperature")
-    override val temperature: Int,
+    override val temperature: Double,
 //    @ColumnInfo(name = "weatherDescriptions")
 //    override val conditionText: String,
 //    @ColumnInfo(name = "weatherIcons")
 //    override val conditionIconUrl: String,
     @ColumnInfo(name = "windSpeed")
-    override val windSpeed: Int,
+    override val windSpeed: Double,
     @ColumnInfo(name = "windDir")
     override val windDirection: String,
     @ColumnInfo(name = "precip")
-    override val precipitationVolume: Int,
+    override val precipitationVolume: Double,
     @ColumnInfo(name = "feelslike")
-    override val feelsLikeTemperature: Int,
+    override val feelsLikeTemperature: Double,
     @ColumnInfo(name = "visibility")
-    override val visibilityDistance: Int
+    override val visibilityDistance: Double
 ) : UnitSpecificCurrentWeatherEntry
