+++ app/src/main/java/com/nakano/forecastmvvm/data/db/unitlocalized/UnitSpecificCurrentWeatherEntry.kt
@@ -1,12 +1,12 @@
 package com.nakano.forecastmvvm.data.db.unitlocalized
 
 interface UnitSpecificCurrentWeatherEntry {
-    val temperature: Int
+    val temperature: Double
 //    val conditionText: String
 //    val conditionIconUrl: String
-    val windSpeed: Int
+    val windSpeed: Double
     val windDirection: String
-    val precipitationVolume: Int
-    val feelsLikeTemperature: Int
-    val visibilityDistance: Int
+    val precipitationVolume: Double
+    val feelsLikeTemperature: Double
+    val visibilityDistance: Double
 }
